07_ops/scripts/requirements.txt
# File extraction dependencies for project brain
# Install with: pip install -r requirements.txt

# PDF extraction
pypdf>=4.0.0

# Word documents (.docx)
python-docx>=1.1.0

# PowerPoint (.pptx)
python-pptx>=0.6.23

# OCR for scanned documents (requires Tesseract installed on system)
pytesseract>=0.3.10
Pillow>=10.0.0

# PDF to image conversion for OCR fallback
pdf2image>=1.16.3

